Move adding events to map into mapUtils
